+++ changelog.md
@@ -6,6 +6,7 @@
 
 ### v1.0.1
 - 修复了造石机和流体提取器kubejs适配错误的bug
+- 修复了龙之套具伤害异常的bug
 
 ## en_us
 ### v1.0.0
@@ -13,3 +14,4 @@
 
 ### v1.0.1
 - Fixed the bug of kubejs compact error for the stone work generator and fluid extractor
+- Fixed the bug of abnormal damage caused by the Dragon Star Kit